Support for new model of DVM.

+++ netconf/src/ncx/val.h
@@ -205,9 +205,9 @@ extern "C" {
 
 #define VAL_INT(V)     ((V)->v.num.i)
 
-#define VAL_INT8(V)    ((int8)((V)->v.num.i))
+#define VAL_INT8(V)    ((V)->v.num.i)
 
-#define VAL_INT16(V)   ((int16)((V)->v.num.i))
+#define VAL_INT16(V)   ((V)->v.num.i)
 
 #define VAL_INT32(V)   ((V)->v.num.i)
 
@@ -225,9 +225,9 @@ extern "C" {
 
 #define VAL_UINT(V)    ((V)->v.num.u)
 
-#define VAL_UINT8(V)    ((uint8)((V)->v.num.u))
+#define VAL_UINT8(V)    ((V)->v.num.u)
 
-#define VAL_UINT16(V)   ((uint16)((V)->v.num.u))
+#define VAL_UINT16(V)   ((V)->v.num.u)
 
 #define VAL_UINT32(V)   ((V)->v.num.u)
